Removed all occurrences of "cast"-function  getInt,-Uint,-Float,-Double,-Double80 Rational,-Uint64,-Int64,-Uint128,-Int128 and replaced them with real case- operators declared as members explicit operator <type>() const... discovered the use of <explicit> cast operators
Modified Register-classes in project IntelCPU, to assure the really are
singletons...and removed virtual functions, and added type- and size-field
to base class instead

+++ Test/UnitTest/TestRandom/TestRandom.cpp
@@ -66,7 +66,7 @@ namespace TestRandom {
         _int128 t128 = b[i];
         t128 *= bucketCount;
         t128 /= length;
-        const UINT index = t128;
+        const UINT index = (UINT)t128;
         observed[index]++;
       }
       const double e = (double)n / bucketCount;
